Show / hide item labels and "are you sure" before delete

+++ components/panes/actions-action.tsx
@@ -135,9 +135,11 @@ const RightAction: FunctionComponent<{
               icon={faTrashAlt}
               size="sm"
               onClick={(e) => {
-                dispatch(deleteActionByUuid(action.uuid));
-                setEditMode(true);
-                e.stopPropagation();
+                if (window.confirm("Are you sure you want to delete this Action?")) {
+                  dispatch(deleteActionByUuid(action.uuid));
+                  setEditMode(true);
+                  e.stopPropagation();
+                }
               }}
             />
           </div>
